Adding README.md, modifying mutation, changing htmlBody to richTextBody to make it generic
