+++ plugins/talk-plugin-rich-text-pell/client/containers/CommentContent.js
@@ -6,7 +6,7 @@ export default withFragments({
   comment: gql`
     fragment TalkPluginRTE_CommentContent_comment on Comment {
       body
-      htmlBody
+      richTextBody
     }
   `,
 })(CommentContent);
